refactor: move everything to src
