+++ src/components/Degrade.js
