+++ src/components/Slide.js
