+++ src/components/degreda.module.scss
