+++ src/components/slide.module.scss
