+++ src/components/slides/SN1.js
