+++ src/components/slides/SN1.module.scss
