+++ src/components/slides/SN10.js
