+++ src/components/slides/SN10.module.scss
