+++ src/components/slides/SN11.js
