+++ src/components/slides/SN12.module.scss
