+++ src/components/slides/SN14.module.scss
