+++ src/components/slides/SN2.module.scss
