+++ src/components/slides/SN3.js
