+++ src/components/slides/SN4.module.scss
