+++ src/components/slides/SN5.js
