+++ src/components/slides/SN5.module.scss
