+++ src/components/slides/SN6.js
