+++ src/components/slides/SN6.module.scss
