+++ src/components/slides/SN7.js
