+++ src/components/slides/SN7.module.scss
