+++ src/components/slides/SN8.js
