+++ src/components/slides/SN8.module.scss
