+++ src/components/slides/SN9.js
