+++ src/components/slides/SN9.module.scss
