+++ src/components/slides/export.js
